requirements.txt
openai>=1.12.0
google-generativeai>=0.3.2
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
Pillow>=10.2.0
python-facebook-api>=0.1.0
selenium>=4.18.0
webdriver-manager>=4.0.1
pytesseract>=0.3.10
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1
python-multipart>=0.0.9 